Add mvvm, proxy, coordinator
