+++ Movies_MVVM/Core/Builder/Protocols/AssemblyModuleBuilderProtocol.swift
@@ -1,0 +1,11 @@
+// AssemblyModuleBuilderProtocol.swift
+// Copyright © RoadMap. All rights reserved.
+
+import UIKit
+
+/// Screens builder protocol
+protocol AssemblyModuleBuilderProtocol {
+    func makeMovieListModule() -> UIViewController
+    func makeDetailMoviesModule(id: String?) -> UIViewController
+    func makeFavoriteMoviesModule() -> UIViewController
+}
