+++ Movies_MVVM/Core/ImageService/Protocols/ProxyProtocol.swift
@@ -1,0 +1,9 @@
+// ProxyProtocol.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// Proxy protocol
+protocol ProxyProtocol {
+    func getImage(imagePath: String, completion: @escaping (Result<Data, Error>) -> ())
+}
