--- Movies_MVVM/Helpers/PublicConstants.swift
@@ -1,9 +1,0 @@
-// PublicConstants.swift
-// Copyright © RoadMap. All rights reserved.
-
-import Foundation
-
-/// Public Constants.
-enum PublicConstants {
-    static let apiKeyText = "7305af56f5e3e8404ad79e2f0c7cefe0"
-}
