+++ Movies_MVVM/Models/Cast/CastResult.swift
@@ -1,0 +1,12 @@
+// CastResult.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// CastResult model.
+struct CastResult: Decodable {
+    /// Film ID
+    let id: Int
+    /// Cast results
+    let cast: [Cast]
+}
