+++ Movies_MVVM/Models/Movies/MovieResult.swift
@@ -1,0 +1,10 @@
+// MovieResult.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// Movie results
+struct MovieResults: Decodable {
+    /// Movie results
+    let results: [Movie]
+}
