--- Movies_MVVM/Modules/ProfileTab/Controllers/FavoriteViewController.swift
@@ -1,7 +1,0 @@
-// FavoriteViewController.swift
-// Copyright © RoadMap. All rights reserved.
-
-import UIKit
-
-/// Favourite movies screen.
-final class FavoriteViewController: UIViewController {}
