+++ Movies_MVVM/Modules/ProfileTab/View/FavoriteViewController.swift
@@ -1,0 +1,7 @@
+// FavoriteViewController.swift
+// Copyright © RoadMap. All rights reserved.
+
+import UIKit
+
+/// Favourite movies screen
+final class FavoriteViewController: UIViewController {}
